Update Install and Import Module.ps1
Without the Error Action, the Try Statement presents an error when the module is missing that does not trigger the catch clause.  As a result, it rolls into the Import-Module directly after the get attempt and the whole script fails if the module was not previously installed.

Try Catch requires the try statement to create a terminating statement in order to engage the catch clause. Apparently the Get-InstalledModule failure is not a terminating error.
https://docs.microsoft.com/en-us/powershell/module/microsoft.powershell.core/about/about_try_catch_finally?view=powershell-7

I tested the original script in my environment, and had errors across all machines.  I added in the error action and success.